backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
psycopg2-binary>=2.9.9
pydicom==1.4.2
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv==1.0.0
nibabel>=5.0.0
dicom2nifti==2.2.8
reportlab>=3.6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
torch>=2.0.0
requests>=2.31.0
PyPDF2>=3.0.0
pypdf>=3.17.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-huggingface>=0.0.1
chromadb>=0.4.22
tiktoken>=0.5.2
pydantic>=2.0.0
opencv-python
ultralytics>=8.0.0
tavily-python